backend/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.1
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
email-validator==2.1.1
